+++ packages/@webex/plugin-meetings/test/unit/spec/multistream/mediaRequestManager.ts
@@ -5,6 +5,8 @@ import {assert} from '@webex/test-helper-chai';
 import {getMaxFs} from '@webex/plugin-meetings/src/multistream/remoteMedia';
 import FakeTimers from '@sinonjs/fake-timers';
 import * as mediaCore from '@webex/internal-media-core';
+import { assertFunction } from '@babel/types';
+import { expect } from 'chai';
 
 type ExpectedActiveSpeaker = {
   policy: 'active-speaker';
@@ -79,7 +81,7 @@ describe('MediaRequestManager', () => {
   });
 
   // helper function for adding an active speaker request
-  const addActiveSpeakerRequest = (priority, receiveSlots, maxFs, commit = false) =>
+  const addActiveSpeakerRequest = (priority, receiveSlots, maxFs, commit = false) => 
     mediaRequestManager.addRequest(
       {
         policyInfo: {
@@ -379,6 +381,26 @@ describe('MediaRequestManager', () => {
         maxMbps: MAX_MBPS_720p,
       },
     ]);
+  });
+
+  it('removes the events maxFsUpdate and sourceUpdate when cancelRequest() is called', async () => {
+
+    const requestId = addActiveSpeakerRequest(255, [fakeReceiveSlots[2], fakeReceiveSlots[3]], MAX_FS_720p);
+
+    mediaRequestManager.cancelRequest(requestId, true);
+
+    const sourceUpdateHandler = fakeReceiveSlots[2].off.getCall(0);
+
+    const maxFsHandlerCall = fakeReceiveSlots[2].off.getCall(1);
+
+    const maxFsEventName = maxFsHandlerCall.args[0];
+    const sourceUpdateEventName = sourceUpdateHandler.args[0];
+
+    expect(sourceUpdateHandler.args[1]).to.be.a('function');
+    expect(maxFsHandlerCall.args[1]).to.be.a('function');
+
+    assert.equal(maxFsEventName, 'maxFsUpdate')
+    assert.equal(sourceUpdateEventName, 'sourceUpdate')
   });
 
   it('cancels the requests correctly when cancelRequest() is called with commit=true', () => {
@@ -1058,3 +1080,7 @@ describe('MediaRequestManager', () => {
     });
   });
 });
+function assertEqual(arg0: any, arg1: string) {
+  throw new Error('Function not implemented.');
+}
+